refactor(tools): extract global-install script into tools/bun-install
Split the monolithic 713-line dev-bun-install-global-command.ts into a proper tools/bun-install/ package with domain modules (types, utils, workspace, archive, operations) and replace the original with a 2-line shim conforming to the project convention (scripts/ as entrypoint shims importing from tools/).
